Replacement of generic thermostat
