+++ homeassistant/components/generic_thermostat.old/manifest.json
@@ -1,0 +1,11 @@
+{
+  "domain": "generic_thermostat",
+  "name": "Generic thermostat",
+  "documentation": "https://www.home-assistant.io/integrations/generic_thermostat",
+  "requirements": [],
+  "dependencies": [
+    "sensor",
+    "switch"
+  ],
+  "codeowners": []
+}
