+++ homeassistant/components/generic_thermostat/manifest.json
@@ -1,8 +1,13 @@
 {
-  "domain": "generic_thermostat",
-  "name": "Generic Thermostat",
-  "documentation": "https://www.home-assistant.io/integrations/generic_thermostat",
-  "requirements": [],
-  "dependencies": ["sensor", "switch"],
-  "codeowners": []
+    "domain": "generic_thermostat",
+    "name": "Generic thermostat",
+    "documentation": "https://www.home-assistant.io/components/generic_thermostat",
+    "requirements": [],
+    "dependencies": [
+        "sensor",
+        "switch"
+    ],
+    "codeowners": [
+        "@DB-CL"
+    ]
 }
