--- homeassistant/components/virtual_thermostat/__init__.py
@@ -1,1 +1,0 @@
-"""The virtual_thermostat component."""
