--- homeassistant/components/virtual_thermostat/manifest.json
@@ -1,13 +1,0 @@
-{
-    "domain": "virtual_thermostat",
-    "name": "Virtual thermostat",
-    "documentation": "https://www.home-assistant.io/components/virtual_thermostat",
-    "requirements": [],
-    "dependencies": [
-        "sensor",
-        "switch"
-    ],
-    "codeowners": [
-        "@DB-CL"
-    ]
-}
