+++ tests/components/generic_thermostat.old/test_climate.py
